refactor: corrige les erreurs clippy (-D warnings)
- clone_on_copy / needless_borrow dans db_interface, les middlewares,
  does_user_exist_by_email et les tests de requêtes.
- Default pour MockEmailClient (new_without_default).
- Validation JWT construite avec leeway: 0 dans l'initialiseur.
- strip_prefix("Bearer ") dans get_jwt_from_request.
- Redis::delete : turbofish explicite au lieu d'un binding de type ().
- smart_db/smart_db.rs renommé en smart_database.rs (module_inception) ;
  l'API publique smart_db::SmartDatabase est inchangée.

+++ tests/queries_tests.rs
@@ -290,7 +290,7 @@ mod queries_tests {
                     VALUES (50, $1, 'Test ACL Group') \
                     ON CONFLICT (id) DO NOTHING",
             )
-            .bind(alice_id as i32)
+            .bind(alice_id)
             .execute(&pool)
             .await
             .unwrap();
@@ -302,7 +302,7 @@ mod queries_tests {
                     (SELECT id FROM public.permissions WHERE action = 'read' AND resource_id = (SELECT id FROM public.resources WHERE name = 'groups') LIMIT 1)) \
                     ON CONFLICT DO NOTHING"
                 )
-                    .bind(alice_id as i32)
+                    .bind(alice_id)
                     .execute(&pool)
                     .await
                     .unwrap();
@@ -344,7 +344,7 @@ mod queries_tests {
                              VALUES (10, $1, 'Confidential Group') \
                              ON CONFLICT (id) DO NOTHING",
             )
-            .bind(alice_id as i32)
+            .bind(alice_id)
             .execute(&pool)
             .await
             .unwrap();
